+++ Tortuga.Chain/Tortuga.Chain.Core/DataSources/ISupportsDisableIndexes.cs
@@ -1,0 +1,30 @@
+namespace Tortuga.Chain.DataSources;
+
+/// <summary>
+/// Used to mark datasources that support enabling and disabling indexes on a table.
+/// </summary>
+public interface ISupportsDisableIndexes
+{
+	/// <summary>
+	/// Disables all of the indexes on the indicated table..
+	/// </summary>
+	/// <param name="tableName">Name of the table.</param>
+	public ILink<int?> DisableIndexes(string tableName);
+
+	/// <summary>
+	/// Enables all of the indexes on the indicated table.
+	/// </summary>
+	/// <param name="tableName">Name of the table.</param>
+
+	public ILink<int?> EnableIndexes(string tableName);
+
+	/// <summary>
+	/// Disables all of the indexes on the indicated table..
+	/// </summary>
+	public ILink<int?> DisableIndexes<TObject>() where TObject : class;
+
+	/// <summary>
+	/// Enables all of the indexes on the indicated table.
+	/// </summary>
+	public ILink<int?> EnableIndexes<TObject>() where TObject : class;
+}